Filter and consolidate reviews logic
