+++ src/components/listing-detail.js
@@ -17,7 +17,7 @@ import {
 
 import { PendingBadge, SoldBadge, FeaturedBadge } from 'components/badges'
 import Modal from 'components/modal'
-import Review from 'components/review'
+import Reviews from 'components/reviews'
 import UserCard from 'components/user-card'
 import { MetamaskModal, ProcessingModal } from 'components/modals/wait-modals'
 
@@ -53,7 +53,6 @@ class ListingsDetail extends Component {
       offers: [],
       pictures: [],
       purchases: [],
-      reviews: [],
       step: this.STEP.VIEW,
       boostLevel: null,
       boostValue: 0,
@@ -76,7 +75,6 @@ class ListingsDetail extends Component {
       // Load from IPFS
       await this.loadListing()
       await this.loadOffers()
-      await this.loadReviews()
     } else if (this.props.listingJson) {
       const obj = Object.assign({}, this.props.listingJson, { loading: false })
       // Listing json passed in directly
@@ -222,20 +220,6 @@ class ListingsDetail extends Component {
     }
   }
 
-  async loadReviews() {
-    try {
-      const reviews = await origin.marketplace.getListingReviews(
-        this.props.listingId
-      )
-      this.setState({
-        reviews: reviews.filter(r => r.reviewer !== this.state.seller)
-      })
-    } catch (error) {
-      console.error(error)
-      console.error(`Error fetching reviews`)
-    }
-  }
-
   resetToStepOne() {
     this.setState({ step: this.STEP.VIEW })
   }
@@ -253,7 +237,6 @@ class ListingsDetail extends Component {
       offers,
       pictures,
       price,
-      reviews,
       seller,
       step
       // unitsRemaining
@@ -705,21 +688,9 @@ class ListingsDetail extends Component {
             <div className="row">
               <div className="col-12 col-md-8">
                 <hr />
-                <div className="reviews">
-                  <h2>
-                    <FormattedMessage
-                      id={'listing-detail.reviews'}
-                      defaultMessage={'Reviews'}
-                    />
-                    &nbsp;
-                    <span className="review-count">
-                      <FormattedNumber value={reviews.length} />
-                    </span>
-                  </h2>
-                  {reviews.map(r => <Review key={r.id} review={r} />)}
-                  {/* To Do: pagination */}
-                  {/* <a href="#" className="reviews-link">Read More<img src="/images/carat-blue.svg" className="down carat" alt="down carat" /></a> */}
-                </div>
+                {this.state.seller && (
+                  <Reviews userAddress={this.state.seller} />
+                )}
               </div>
             </div>
           )}
