+++ src/components/purchase-detail.js
@@ -26,7 +26,7 @@ import { RejectionModal, WithdrawModal } from 'components/modals/offer-modals'
 import { MetamaskModal } from 'components/modals/wait-modals'
 import OfferStatusEvent from 'components/offer-status-event'
 import PurchaseProgress from 'components/purchase-progress'
-import Review from 'components/review'
+import Reviews from 'components/reviews'
 import TransactionHistory from 'components/transaction-history'
 import UnnamedUser from 'components/unnamed-user'
 import UserCard from 'components/user-card'
@@ -60,7 +60,6 @@ const defaultState = {
   problemInferred: false,
   processing: false,
   purchase: {},
-  reviews: [],
   seller: {},
   areSellerStepsOpen: true
 }
@@ -309,11 +308,9 @@ class PurchaseDetail extends Component {
     try {
       const purchase = await origin.marketplace.getOffer(offerId)
       const listing = await getListing(purchase.listingId, true)
-      const reviews = await origin.marketplace.getListingReviews(offerId)
       this.setState({
         listing,
-        purchase,
-        reviews
+        purchase
       })
       if (listing) {
         this.getListingSchema()
@@ -671,7 +668,6 @@ class PurchaseDetail extends Component {
       problemInferred,
       processing,
       purchase,
-      reviews,
       seller,
       translatedSchema,
       areSellerStepsOpen
@@ -1141,22 +1137,9 @@ class PurchaseDetail extends Component {
                     )}
                   </div>
                   <hr />
+                  <Reviews userAddress={listing.seller} />
                 </Fragment>
               )}
-              <div className="reviews">
-                <h2>
-                  <FormattedMessage
-                    id={'purchase-detail.reviewsHeading'}
-                    defaultMessage={'Reviews'}
-                  />
-                  &nbsp;<span className="review-count">
-                    {Number(reviews.length).toLocaleString()}
-                  </span>
-                </h2>
-                {reviews.map(r => <Review key={r.id} review={r} />)}
-                {/* To Do: pagination */}
-                {/* <a href="#" className="reviews-link">Read More<img src="/images/carat-blue.svg" className="down carat" alt="down carat" /></a> */}
-              </div>
             </div>
             <div className="col-12 col-lg-4">
               {created && (
