+++ translations/all-messages.json
@@ -184,7 +184,6 @@
   "listing-detail.sellerSold": "You've sold this listing.",
   "listing-detail.viewOffer": "View Offer",
   "listing-detail.viewSale": "View Sale",
-  "listing-detail.reviews": "Reviews",
   "listings-grid.originContractNotFound": "No Origin listing contracts were found on this network.",
   "listings-grid.changeNetworks": "You may need to change networks, or deploy the contract.",
   "listings-grid.listingsCount": "{listingIdsCount} Listings",
@@ -361,7 +360,6 @@
   "purchase-detail.transactionHistoryHeading": "Transaction History",
   "purchase-detail.listingDetails": "Listing Details",
   "purchase-detail.viewOnIPFS": "View on IPFS",
-  "purchase-detail.reviewsHeading": "Reviews",
   "purchase-detail.summaryPending": "Pending",
   "purchase-detail.summarySold": "Sold",
   "purchase-detail.price": "Price",
@@ -373,6 +371,7 @@
   "purchase-progress.disputedStarted": "Dispute Started",
   "purchase-progress.rulingMade": "Ruling Made",
   "review.unnamedUser": "Unnamed User",
+  "User.reviews": "Reviews",
   "SearchResults.filterGroup.searchFiltersClear": "Clear",
   "SearchResults.filterGroup.searchFiltersApply": "Apply",
   "SearchResults.searchFiltersClear": "Clear",
@@ -555,7 +554,6 @@
   "User.facebook": "Facebook",
   "User.twitter": "Twitter",
   "User.airbnb": "Airbnb",
-  "User.reviews": "Reviews",
   "schema.announcements.title": "Title",
   "schema.announcements.examples": "Examples",
   "schema.announcements.category": "Category",
