+++ translations/messages/src/components/listing-detail.json
@@ -118,9 +118,5 @@
   {
     "id": "listing-detail.viewSale",
     "defaultMessage": "View Sale"
-  },
-  {
-    "id": "listing-detail.reviews",
-    "defaultMessage": "Reviews"
   }
 ]
