+++ translations/messages/src/components/reviews.json
@@ -1,0 +1,6 @@
+[
+  {
+    "id": "User.reviews",
+    "defaultMessage": "Reviews"
+  }
+]
